[OF-1410] fix: fixed http request crash
