+++ Library/include/CSP/Multiplayer/Components/HotspotSpaceComponent.h
@@ -111,9 +111,6 @@ public:
 	void SetIsARVisible(bool InValue) override;
 	/// @}
 
-	/// @brief Called internally when the hotspot is deleted
-	CSP_NO_EXPORT void RemoveSequences(csp::systems::NullResultCallback Callback);
-
 private:
 	void OnLocalDelete() override;
 };
